some minor fixes
svn path=/trunk/kdesupport/emerge/; revision=1108731

+++ portage/win32libs-sources/zlib-src/zlib-src-1.2.4.py
@@ -8,7 +8,7 @@ class subinfo(info.infoclass):
        for ver in [ '1.2.4' ]:
            self.targets[ ver ] = 'http://downloads.sourceforge.net/sourceforge/libpng/zlib-%s.tar.gz' % ver
            self.targetInstSrc[ ver ] = "zlib-" + ver
-       self.patchToApply['1.2.4'] = ("zlib-src-1.2.4-20100329.diff", 1)
+       self.patchToApply['1.2.4'] = ("zlib-1.2.4-20100329.diff", 1)
        self.defaultTarget = '1.2.4'
 
     def setDependencies( self ):
@@ -23,7 +23,8 @@ class Package(CMakePackageBase):
     def unpack(self):
         if not CMakePackageBase.unpack( self ):
             return False       
-        os.remove(os.path.join( self.sourceDir() , "zconf.h" ))
+        if os.path.exists(os.path.join( self.sourceDir() , "zconf.h" )):
+            os.remove(os.path.join( self.sourceDir() , "zconf.h" ))
         return True
 
     def createPackage( self ):
